+++ src/neuropixel.py
@@ -337,9 +337,9 @@ class NP2Converter:
         self.sos_lp = scipy.signal.butter(**butter_lp_kwargs, output="sos")
 
         # Number of ap channels
-        self.napch = int(self.sr.meta["snsApLfSy"][0])
+        self.napch = self.sr.nc - self.sr.nsync
         # Position of start of sync channels in the raw data
-        self.idxsyncch = int(self.sr.meta["snsApLfSy"][0])
+        self.idxsyncch = np.arange(self.sr.nsync) + self.sr.nc - self.sr.nsync
 
         self.extra = extra or ""
         self.nshank = nshank or None
@@ -372,8 +372,12 @@ class NP2Converter:
             status = self._process_NP21(overwrite=overwrite)
         elif self.np_version == "NP2.1":
             status = self._process_NP21(overwrite=overwrite)
+        elif self.np_version == "NP2QB":
+            status = self._process_NP24(overwrite=overwrite)
         else:
-            _logger.warning("Meta file is not of type NP2.1 or NP2.4, cannot process")
+            _logger.warning(
+                f"Probe version {self.np_version} unknown. Should be NP2.1, NP2.4 or NP2QB."
+            )
             status = -1
         return status
 
@@ -406,11 +410,9 @@ class NP2Converter:
 
         for first, last in wg.firstlast:
             chunk_ap = self.sr._raw[first:last, : self.napch].T
-            chunk_ap_sync = self.sr._raw[first:last, self.idxsyncch :].T
+            chunk_ap_sync = self.sr._raw[first:last, self.idxsyncch].T
             chunk_lf = self.extract_lfp(self.sr[first:last, : self.napch].T)
-            chunk_lf_sync = self.extract_lfp_sync(
-                self.sr[first:last, self.idxsyncch :].T
-            )
+            chunk_lf_sync = self.extract_lfp_sync(self.sr[first:last, self.idxsyncch].T)
 
             chunk_ap2save = self._ind2save(
                 chunk_ap, chunk_ap_sync, wg, ratio=1, etype="ap"
@@ -425,8 +427,8 @@ class NP2Converter:
         self._closefiles(etype="ap")
         self._closefiles(etype="lf")
 
-        self._writemetadata_ap()
-        self._writemetadata_lf()
+        self._writemetadata(etype="ap")
+        self._writemetadata(etype="lf")
 
         if self.post_check:
             self.check_NP24()
@@ -453,14 +455,23 @@ class NP2Converter:
         shank_info = {}
         self.already_exists = False
 
-        for sh in n_shanks:
+        for i, sh in enumerate(n_shanks):
             _shank_info = {}
             # channels for individual shank + sync channel
+            # NP2QB has a sync channel per shank
+            if self.np_version == "NP2.4":
+                isync = self.idxsyncch
+            elif self.np_version == "NP2QB":
+                isync = self.idxsyncch[i]
+            else:
+                raise ValueError(
+                    f"Probe version {self.np_version} should be 'NP2.4' or 'NP2QB'"
+                )
             _shank_info["chns"] = np.r_[
                 np.where(chn_info["shank"] == sh)[0],
-                np.array(spikeglx._get_sync_trace_indices_from_meta(self.sr.meta)),
+                isync,
             ]
-
+            # we name the probe folder by appending a, b, c etc..
             probe_path = self.ap_file.parent.parent.joinpath(
                 label + chr(97 + int(sh)) + self.extra
             )
@@ -513,9 +524,7 @@ class NP2Converter:
             last = last + offset
 
             chunk_lf = self.extract_lfp(self.sr[first:last, : self.napch].T)
-            chunk_lf_sync = self.extract_lfp_sync(
-                self.sr[first:last, self.idxsyncch :].T
-            )
+            chunk_lf_sync = self.extract_lfp_sync(self.sr[first:last, self.idxsyncch].T)
 
             chunk_lf2save = self._ind2save(
                 chunk_lf, chunk_lf_sync, wg, ratio=self.ratio, etype="lf"
@@ -525,7 +534,7 @@ class NP2Converter:
 
         self._closefiles(etype="lf")
 
-        self._writemetadata_lf()
+        self._writemetadata(etype="lf")
 
         if self.compress:
             self.compress_NP21(overwrite=overwrite)
@@ -599,19 +608,12 @@ class NP2Converter:
             chunk = np.zeros_like(expected)
             for ish, sh in enumerate(self.shank_info.keys()):
                 srs = self.shank_info[sh]["sr"]
-                if ish == 0:
-                    chunk[:, self.shank_info[sh]["chns"]] = srs[first:last, :]
-                else:
-                    chunk[:, self.shank_info[sh]["chns"][:-1]] = srs[first:last, :-1]
-            assert np.array_equal(expected, chunk), (
-                "data in original file and split files do no match"
-            )
-
+                chunk[:, self.shank_info[sh]["chns"]] = srs[first:last, :]
+            np.testing.assert_array_equal(expected, chunk)
         # close the sglx instances once we are done checking
         for sh in self.shank_info.keys():
             sr = self.shank_info[sh].pop("sr")
             sr.close()
-
         self.check_completed = True
 
     def compress_NP24(self, overwrite=False, **kwargs):
@@ -681,7 +683,7 @@ class NP2Converter:
 
     def _split2shanks(self, chunk, etype="ap"):
         """
-        Splits the signal on the 384 channels into the individual shanks and saves to file
+        Splits the signal on the full 384 / 1536 channels into the individual shanks and saves to file
 
         :param chunk: portion of signal with all 384 channels
         :param type: ephys type, either 'ap' or 'lf'
@@ -720,7 +722,7 @@ class NP2Converter:
                     chunk[:, slice(*ind2save)].T
                     / self.sr.channel_conversion_sample2v[etype][: self.napch],
                     chunk_sync[:, slice(*ind2save)].T
-                    / self.sr.channel_conversion_sample2v[etype][self.idxsyncch :],
+                    / self.sr.channel_conversion_sample2v[etype][self.idxsyncch],
                 ]
             ).astype(np.int16)
         else:
@@ -769,33 +771,7 @@ class NP2Converter:
             open = self.shank_info[sh].pop(f"{etype}_open_file")
             open.close()
 
-    def _writemetadata_ap(self):
-        """
-        Function to create ap meta data file. Adapts the relevant keys in the spikeglx meta file
-        to contain the correct number of channels. Also adds key to indicate that this is not an
-        original meta data file, but one that has been adapted
-
-        :return:
-        """
-
-        for sh in self.shank_info.keys():
-            n_chns = len(self.shank_info[sh]["chns"])
-            # First for the ap file
-            meta_shank = copy.deepcopy(self.sr.meta)
-            meta_shank["acqApLfSy"][0] = n_chns - 1
-            meta_shank["snsApLfSy"][0] = n_chns - 1
-            meta_shank["nSavedChans"] = n_chns
-            meta_shank["fileSizeBytes"] = self.shank_info[sh]["ap_file"].stat().st_size
-            meta_shank["snsSaveChanSubset_orig"] = spikeglx._get_savedChans_subset(
-                self.shank_info[sh]["chns"]
-            )
-            meta_shank["snsSaveChanSubset"] = f"0:{n_chns - 1}"
-            meta_shank["original_meta"] = False
-            meta_shank[f"{self.np_version}_shank"] = int(sh[-1])
-            meta_file = self.shank_info[sh]["ap_file"].with_suffix(".meta")
-            spikeglx.write_meta_data(meta_shank, meta_file)
-
-    def _writemetadata_lf(self):
+    def _writemetadata(self, etype="ap"):
         """
         Function to create lf meta data file. Adapts the relevant keys in the spikeglx meta file
         to contain the correct number of channels. Also adds key to indicate that this is not an
@@ -803,25 +779,34 @@ class NP2Converter:
 
         :return:
         """
-
+        if etype == "ap":
+            ifull, iempty, fkey, fs = (0, 1, "ap_file", self.sr.fs)  # ap
+        elif etype == "lf":
+            ifull, iempty, fkey, fs = (1, 0, "lf_file", self.fs_lf)  # lf
+        else:
+            ValueError(f"Unsupported etype: {etype}")
         for sh in self.shank_info.keys():
             n_chns = len(self.shank_info[sh]["chns"])
             meta_shank = copy.deepcopy(self.sr.meta)
-            meta_shank["acqApLfSy"][0] = 0
-            meta_shank["acqApLfSy"][1] = n_chns - 1
-            meta_shank["snsApLfSy"][0] = 0
-            meta_shank["snsApLfSy"][1] = n_chns - 1
-            meta_shank["fileSizeBytes"] = self.shank_info[sh]["lf_file"].stat().st_size
-            meta_shank["imSampRate"] = self.fs_lf
-            if self.np_version == "NP2.4":
+            meta_shank["acqApLfSy"][iempty] = 0
+            meta_shank["acqApLfSy"][ifull] = n_chns - 1
+            meta_shank["acqApLfSy"][2] = 1  # for NP2QB this goes from 4 to 1
+            meta_shank["snsApLfSy"][iempty] = 0
+            meta_shank["snsApLfSy"][ifull] = n_chns - 1
+            meta_shank["snsApLfSy"][2] = 1  # for NP2QB this goes from 4 to 1
+            meta_shank["fileSizeBytes"] = self.shank_info[sh][fkey].stat().st_size
+            meta_shank["imSampRate"] = fs
+            if self.np_version in ["NP2.4", "NP2QB"]:
                 meta_shank["snsSaveChanSubset_orig"] = spikeglx._get_savedChans_subset(
                     self.shank_info[sh]["chns"]
                 )
                 meta_shank["snsSaveChanSubset"] = f"0:{n_chns - 1}"
                 meta_shank["nSavedChans"] = n_chns
+                meta_shank["NP2.4_shank"] = int(sh[-1])
+            else:
+                meta_shank[f"{self.np_version}_shank"] = int(sh[-1])
             meta_shank["original_meta"] = False
-            meta_shank[f"{self.np_version}_shank"] = int(sh[-1])
-            meta_file = self.shank_info[sh]["lf_file"].with_suffix(".meta")
+            meta_file = self.shank_info[sh][fkey].with_suffix(".meta")
             spikeglx.write_meta_data(meta_shank, meta_file)
 
     def get_processed_files_NP24(self):